Add error handling and model validation

+++ GreenDemic/Controllers/ItemController.cs
@@ -15,20 +15,23 @@ namespace GreenDemic.Controllers
 {
     public class ItemController : Controller
     {
-        private readonly ILogger<ItemController> _logger;
+        // DAL context
         private ItemDAL itemContext = new ItemDAL();
         private ShoppingBagItemDAL shoppingBagItemContext = new ShoppingBagItemDAL();
         private ShoppingBagDAL shoppingBagContext = new ShoppingBagDAL();
+
+        // Logging 
+        private readonly ILogger<ItemController> _logger;
         public ItemController(ILogger<ItemController> logger)
         {
             _logger = logger;
         }
 
-        // Return area interest list for dropdown
+        // Return category of food for dropdown list
         private List<SelectListItem> GetCategory()
         {
             List<SelectListItem> categoryList = new List<SelectListItem>();
-            List<String> myCategoryList = new List<string>() { "Snack","Meat","Seafood","Dairy","Grains","Fruit","Vegetable","Others" };
+            List<String> myCategoryList = new List<string>() { "Snack", "Meat", "Seafood", "Dairy", "Grains", "Fruit", "Vegetable", "Others" };
 
             foreach (String cat in myCategoryList)
             {
@@ -42,6 +45,7 @@ namespace GreenDemic.Controllers
             return categoryList;
         }
 
+        //Returns list of itemviewmodel based on shopping bag ID
         public List<ItemViewModel> MapToItemVM(int shoppingBagID)
         {
             List<ItemViewModel> itemVMList = new List<ItemViewModel>();
@@ -65,6 +69,7 @@ namespace GreenDemic.Controllers
             return itemVMList;
         }
 
+        //Return itemviewmodel based on a shopping bag and item ID
         public ItemViewModel MapToItemVM(int shoppingBagID, int itemID)
         {
             Item item = itemContext.GetDetails(itemID);
@@ -85,22 +90,43 @@ namespace GreenDemic.Controllers
         // GET: ItemController
         public ActionResult Index(int? id)
         {
-            List<ItemViewModel> itemVMList = MapToItemVM(id.Value);
-            ViewData["ShoppingBagId"] = id.Value;
-            return View(itemVMList);
-        }
+            // Authenticate user
+            if ((HttpContext.Session.GetString("Role") == null) ||
+            (HttpContext.Session.GetString("Role") != "User"))
+            {
+                return RedirectToAction("Index", "Home");
+            }
 
-        // GET: ItemController/Details/5
-        public ActionResult Details(int id)
-        {
-            return View();
+            // If ID is not supplied
+            if (id == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
+            List<ItemViewModel> itemVMList = MapToItemVM(id.Value);
+            ViewData["ShoppingBagId"] = id.Value; //for create and delete item 
+            return View(itemVMList);
         }
 
         // GET: ItemController/Create
         public ActionResult Create(int? id)
         {
+            // Authenticate user
+            if ((HttpContext.Session.GetString("Role") == null) ||
+            (HttpContext.Session.GetString("Role") != "User"))
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
+            // If ID is not supplied
+            if (id == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             ViewData["ShoppingBagId"] = id.Value;
             ViewData["CategoryList"] = GetCategory();
+            ViewData["IsFailed"] = false;
             return View();
         }
 
@@ -109,27 +135,66 @@ namespace GreenDemic.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create(ItemViewModel model, int? id)
         {
-            Item item = new Item();
-            item.ItemID = model.ItemID;
-            item.ItemName = model.ItemName;
-            item.Category = model.Category;
-            item.Cal = model.Cal;
-            int itemID = itemContext.Add(item);
-
-            ShoppingBagItem shoppingBagItem = new ShoppingBagItem();
-            shoppingBagItem.ItemID = itemID;
-            shoppingBagItem.Qty = model.Qty;
-            shoppingBagItem.ShoppingBagID = model.ShoppingBagID;
-            shoppingBagItemContext.Add(shoppingBagItem);
-
+            ViewData["IsFailed"] = false;
             ViewData["ShoppingBagId"] = id.Value;
             ViewData["CategoryList"] = GetCategory();
-            return RedirectToAction("Index","Item", new { id = ViewData["ShoppingBagId"] });
+
+            // Adding new item in shopping bag
+            try
+            {
+                if (ModelState.IsValid)
+                {
+                    Item item = new Item();
+                    item.ItemID = model.ItemID;
+                    item.ItemName = model.ItemName;
+                    item.Category = model.Category;
+                    item.Cal = model.Cal;
+                    int itemID = itemContext.Add(item);
+
+                    ShoppingBagItem shoppingBagItem = new ShoppingBagItem();
+                    shoppingBagItem.ItemID = itemID;
+                    shoppingBagItem.Qty = model.Qty;
+                    shoppingBagItem.ShoppingBagID = model.ShoppingBagID;
+                    shoppingBagItemContext.Add(shoppingBagItem);
+                    return RedirectToAction("Index", "Item", new { id = ViewData["ShoppingBagId"] });
+                }
+                else
+                {
+                    return View(model);
+                }
+
+            }
+            // Shows error message
+            catch
+            {
+                ViewData["IsFailed"] = true;
+                return View(model);
+            }
         }
 
         // GET: ItemController/Edit/5
         public ActionResult Edit(int? id, int? sbID)
         {
+            // Authenticate user
+            if ((HttpContext.Session.GetString("Role") == null) ||
+            (HttpContext.Session.GetString("Role") != "User"))
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
+            // If ID is not supplied
+            if (id == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
+            // If sbID is not supplid
+            if (sbID == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
+            ViewData["IsFailed"] = false;
             ViewData["ShoppingBagId"] = sbID.Value;
             ViewData["CategoryList"] = GetCategory();
             ItemViewModel itemViewModel = MapToItemVM(sbID.Value, id.Value);
@@ -141,28 +206,65 @@ namespace GreenDemic.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit(ItemViewModel itemVM)
         {
-            Item item = new Item();
-            item.ItemID = itemVM.ItemID;
-            item.ItemName = itemVM.ItemName;
-            item.Cal = itemVM.Cal;
-            item.Category = itemVM.Category;
-
-            ShoppingBagItem shoppingBagItem = new ShoppingBagItem();
-            shoppingBagItem.ShoppingBagID = itemVM.ShoppingBagID;
-            shoppingBagItem.ItemID = itemVM.ItemID;
-            shoppingBagItem.Qty = itemVM.Qty;
-
-            itemContext.Update(item, shoppingBagItem);
-
             ViewData["ShoppingBagId"] = itemVM.ShoppingBagID;
             ViewData["CategoryList"] = GetCategory();
-            return RedirectToAction("Index", "Item", new { id = ViewData["ShoppingBagId"] });
+            ViewData["IsFailed"] = false;
+
+            // Update item view model (shopping bag + item)
+            try
+            {
+                if (ModelState.IsValid)
+                {
+                    Item item = new Item();
+                    item.ItemID = itemVM.ItemID;
+                    item.ItemName = itemVM.ItemName;
+                    item.Cal = itemVM.Cal;
+                    item.Category = itemVM.Category;
+
+                    ShoppingBagItem shoppingBagItem = new ShoppingBagItem();
+                    shoppingBagItem.ShoppingBagID = itemVM.ShoppingBagID;
+                    shoppingBagItem.ItemID = itemVM.ItemID;
+                    shoppingBagItem.Qty = itemVM.Qty;
+                    itemContext.Update(item, shoppingBagItem);
+                    return RedirectToAction("Index", "Item", new { id = ViewData["ShoppingBagId"] });
+                }
+                else
+                {
+                    return View(itemVM);
+                }
+            }
+            // Shows error messages
+            catch
+            {
+                ViewData["IsFailed"] = true;
+                return View(itemVM);
+            }
         }
 
         // GET: ItemController/Delete/5
         public ActionResult Delete(int? id, int? sbID)
         {
+            // Authenticate user
+            if ((HttpContext.Session.GetString("Role") == null) ||
+            (HttpContext.Session.GetString("Role") != "User"))
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
+            // If id is not supplied
+            if (id == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
+            // IF sbID is not supplied
+            if (sbID == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             ViewData["ShoppingBagId"] = sbID.Value;
+            ViewData["IsFailed"] = false;
             ItemViewModel itemVM = MapToItemVM(sbID.Value, id.Value);
             return View(itemVM);
         }
@@ -172,8 +274,26 @@ namespace GreenDemic.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Delete(ItemViewModel itemVM)
         {
-            itemContext.Delete(itemVM.ItemID, itemVM.ShoppingBagID);
-            return RedirectToAction("Index", "ShoppingBag");
+            ViewData["IsFailed"] = false;
+            // Delete item from the bag
+            try
+            {
+                if (ModelState.IsValid)
+                {
+                    itemContext.Delete(itemVM.ItemID, itemVM.ShoppingBagID);
+                    return RedirectToAction("Index", "ShoppingBag");
+                }
+                else
+                {
+                    return View(itemVM);
+                }
+            }
+            // Shows error message
+            catch
+            {
+                ViewData["IsFailed"] = true;
+                return View(itemVM);
+            }
         }
     }
 }
