+++ GreenDemic/Controllers/ShoppingBagController.cs
@@ -14,16 +14,19 @@ namespace GreenDemic.Controllers
 {
     public class ShoppingBagController : Controller
     {
-        // Initialise DAL class
+        // DAL context
         private ShoppingBagDAL shoppingBagContext = new ShoppingBagDAL();
-        private readonly ILogger<ShoppingBagController> _logger;
         private ItemDAL itemContext = new ItemDAL();
         private ShoppingBagItemDAL shoppingBagItemContext = new ShoppingBagItemDAL();
 
+        //Logging
+        private readonly ILogger<ShoppingBagController> _logger;
         public ShoppingBagController(ILogger<ShoppingBagController> logger)
         {
             _logger = logger;
         }
+
+        //Calculate total calories in the shopping bag
         private int CalculateCalTotal(List<ItemViewModel> itemVMList)
         {
             int totalCal = 0;
@@ -34,6 +37,8 @@ namespace GreenDemic.Controllers
             }
             return totalCal;
         }
+
+        //Returns list of itemviewmodel based on shopping bag ID
         public List<ItemViewModel> MapToItemVM(int shoppingBagID)
         {
             List<ItemViewModel> itemVMList = new List<ItemViewModel>();
@@ -56,6 +61,8 @@ namespace GreenDemic.Controllers
             }
             return itemVMList;
         }
+
+        //Returns shoppingbagviewmodel based on shopping bag ID
         public ShoppingBagViewModel MapToShoppingBagVM(ShoppingBag shoppingBag)
         {
             List<ItemViewModel> itemVMList = MapToItemVM(shoppingBag.ShoppingBagID);
@@ -70,13 +77,21 @@ namespace GreenDemic.Controllers
 
             return shoppingBagViewModel;
         }
+
         // GET: ShoppingBagController
         public ActionResult Index()
         {
+            // Authenticate user
+            if ((HttpContext.Session.GetString("Role") == null) ||
+            (HttpContext.Session.GetString("Role") != "User"))
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             int accID = (int)HttpContext.Session.GetInt32("AccID");
             List<ShoppingBag> bagList = shoppingBagContext.GetAllShoppingBags(accID);
             List<ShoppingBagViewModel> shoppingBagViewModelList = new List<ShoppingBagViewModel>();
-            foreach(ShoppingBag bag in bagList)
+            foreach (ShoppingBag bag in bagList)
             {
                 ShoppingBagViewModel shoppingBagVM = MapToShoppingBagVM(bag);
                 shoppingBagViewModelList.Add(shoppingBagVM);
@@ -87,6 +102,14 @@ namespace GreenDemic.Controllers
         // GET: ShoppingBagController/Create
         public ActionResult Create()
         {
+            // Authenticate user
+            if ((HttpContext.Session.GetString("Role") == null) ||
+            (HttpContext.Session.GetString("Role") != "User"))
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
+            ViewData["IsFailed"] = false;
             return View();
         }
 
@@ -95,25 +118,45 @@ namespace GreenDemic.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create(ShoppingBag shoppingBag)
         {
+            ViewData["IsFailed"] = false;
+            // Add shopping bag
             try
             {
-                shoppingBag.ShoppingBagID = shoppingBagContext.Add(shoppingBag);
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    shoppingBag.ShoppingBagID = shoppingBagContext.Add(shoppingBag);
+                    return RedirectToAction("Index");
+                }
+                else
+                {
+                    return View(shoppingBag);
+                }
+
             }
+            // Shows error messages
             catch
             {
-                return View();
+                ViewData["IsFailed"] = true;
+                return View(shoppingBag);
             }
         }
 
         // GET: ShoppingBagController/Edit/5
         public ActionResult Edit(int? id)
         {
+            // Authenticate user
+            if ((HttpContext.Session.GetString("Role") == null) ||
+            (HttpContext.Session.GetString("Role") != "User"))
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             if (id == null)
             {
                 return RedirectToAction("Index");
             }
             ShoppingBag bag = shoppingBagContext.GetDetails(id.Value);
+            ViewData["IsFailed"] = false;
             return View(bag);
         }
 
@@ -122,6 +165,7 @@ namespace GreenDemic.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit(ShoppingBag bag)
         {
+            ViewData["IsFailed"] = false;
             try
             {
                 if (ModelState.IsValid)
@@ -136,6 +180,7 @@ namespace GreenDemic.Controllers
             }
             catch
             {
+                ViewData["IsFailed"] = true;
                 return View(bag);
             }
         }
@@ -143,12 +188,20 @@ namespace GreenDemic.Controllers
         // GET: ShoppingBagController/Delete/5
         public ActionResult Delete(int? id)
         {
+            // Authenticate user
+            if ((HttpContext.Session.GetString("Role") == null) ||
+            (HttpContext.Session.GetString("Role") != "User"))
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             if (id == null)
             {
                 return RedirectToAction("Index");
             }
 
             ShoppingBag bag = shoppingBagContext.GetDetails(id.Value);
+            ViewData["IsFailed"] = false;
             return View(bag);
         }
 
@@ -157,9 +210,24 @@ namespace GreenDemic.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Delete(ShoppingBag bag)
         {
-                shoppingBagContext.Delete(bag.ShoppingBagID);
-                return RedirectToAction("Index");
+            ViewData["IsFailed"] = false;
+            try
+            {
+                if (ModelState.IsValid)
+                {
+                    shoppingBagContext.Delete(bag.ShoppingBagID);
+                    return RedirectToAction("Index");
+                }
+                else
+                {
+                    return View(bag);
+                }
+            }
+            catch
+            {
+                ViewData["IsFailed"] = true;
+                return View(bag);
+            }
         }
-
     }
 }
